extract race interstitial-specific functionality from user.rb (and tests from user_test.rb) into standalone helper
