+++ dashboard/test/models/user_test.rb
@@ -2713,71 +2713,6 @@ class UserTest < ActiveSupport::TestCase
     assert student.account_age_days == 10
   end
 
-  def mock_geocoder_result(result)
-    mock_us_object = OpenStruct.new(country_code: result)
-    Geocoder.stubs(:search).returns([mock_us_object])
-  end
-
-  test 'do not show race interstitial to teacher' do
-    mock_geocoder_result('US')
-    teacher = create :teacher, created_at: DateTime.now - 8
-    refute teacher.show_race_interstitial?('ignored_ip')
-  end
-
-  test 'do not show race interstitial to user accounts under 13' do
-    mock_geocoder_result('US')
-    student = User.create(@good_data_young)
-    student.created_at = DateTime.now - 8
-    refute student.show_race_interstitial?('ignored_ip')
-  end
-
-  test 'do not show race interstitial to user accounts less than one week old' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 3
-    refute student.show_race_interstitial?('ignored_ip')
-  end
-
-  test 'do not show race interstitial to user accounts that have already entered race information' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 8
-    student.update_columns(races: 'white,black')
-    refute student.show_race_interstitial?('ignored_ip')
-  end
-
-  test 'do not show race interstitial to user accounts that have closed the dialog already' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 8
-    student.update_columns(races: 'closed_dialog')
-    refute student.show_race_interstitial?('ignored_ip')
-  end
-
-  test 'do not show race interstitial if IP address is nil' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 8
-    mock_ip = nil
-    refute RaceInterstitialHelper.show_race_interstitial?(student, mock_ip)
-  end
-
-  test 'do not show race interstitial to non-US users' do
-    mock_geocoder_result('CA')
-    student = create :student, created_at: DateTime.now - 8
-    unused_ip = 'ignored'
-    refute RaceInterstitialHelper.show_race_interstitial?(student, unused_ip)
-  end
-
-  test 'show race interstitial to US users' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 8
-    unused_ip = 'ignored'
-    assert RaceInterstitialHelper.show_race_interstitial?(student, unused_ip)
-  end
-
-  test 'show race interstitial for student over 13 with account more than 1 week old' do
-    mock_geocoder_result('US')
-    student = create :student, created_at: DateTime.now - 8
-    assert student.show_race_interstitial?('ignored_ip')
-  end
-
   test 'new users must have valid email addresses' do
     assert_creates User do
       create :user, email: 'valid@example.net'
